+++ RELEASES.md
@@ -482,6 +482,12 @@ Examples:
   python build_tools/install_rocm_from_artifacts.py --release 6.4.0.dev0+e015c807437eaf32dac6c14a9c4f752770c51b14 --amdgpu-family gfx110X-dgpu
   ```
 
+- Downloads all gfx94X S3 artifacts from [GitHub CI workflow run 19644138192](https://github.com/ROCm/rocm-libraries/actions/runs/19644138192) in the `ROCm/rocm-libraries` repository:
+
+  ```bash
+  python build_tools/install_rocm_from_artifacts.py --run-id 19644138192 --amdgpu-family gfx94X-dcgpu --tests --run-github-repo ROCm/rocm-libraries
+  ```
+
 Select your AMD GPU family from this file [therock_amdgpu_targets.cmake](https://github.com/ROCm/TheRock/blob/59c324a759e8ccdfe5a56e0ebe72a13ffbc04c1f/cmake/therock_amdgpu_targets.cmake#L44-L81)
 
 By default for CI workflow retrieval, all artifacts (excluding test artifacts) will be downloaded. For specific artifacts, pass in the flag such as `--rand` (RAND artifacts). For test artifacts, pass in the flag `--tests` (test artifacts). For base artifacts only, pass in the flag `--base-only`